toolkit/components/passwordmgr/test/test_basic_form_observer_autofillForms.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: simple form with autofillForms disabled and notifying observers
<script>
commonInit(true);
SimpleTest.waitForExplicitFinish();

const Cc = SpecialPowers.Cc;
const Ci = SpecialPowers.Ci;

// Assume that the pref starts out true, so set to false
SpecialPowers.setBoolPref("signon.autofillForms", false);

var TestObserver = {
  receivedNotificationFoundForm : false,
  receivedNotificationFoundLogins : false,
  dataFoundForm : "",
  dataFoundLogins : null,
  observe : function (subject, topic, data) {
    var pwmgr = Cc["@mozilla.org/login-manager;1"].
                getService(Ci.nsILoginManager);
    if (topic == "passwordmgr-found-form") {
      info("got passwordmgr-found-form");
      this.receivedNotificationFoundForm = true;
      this.dataFoundForm = data;
      // Now fill the form
      pwmgr.fillForm(subject)
           .then(window.startTest)
           .then(null, function(e) { alert(e); });
    } else if (topic == "passwordmgr-found-logins") {
      info("got passwordmgr-found-logins");

      // We only care about the first notification (the second comes from our
      // own call to pwmgr.fillForm.
      if (this.receivedNotificationFoundLogins)
        return;
      this.receivedNotificationFoundLogins = true;
      this.dataFoundLogins = subject.QueryInterface(Ci.nsIPropertyBag2);
    }
  }
};

// Add the observer
SpecialPowers.addObserver(TestObserver, "passwordmgr-found-form", false);
SpecialPowers.addObserver(TestObserver, "passwordmgr-found-logins", false);
</script>

<p id="display"></p>

<div id="content" style="display: block">

  <form id="form1" action="formtest.js">
    <p>This is form 1.</p>
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: simple form with autofillForms disabled and notifying observers **/
function startTest() {
  // Test that found-form observer is notified & got correct data
  is(TestObserver.receivedNotificationFoundForm, true, "Checking found-form observer was notified");
  is(TestObserver.dataFoundForm, "noAutofillForms", "Checking found-form observer got correct data");

  // Check that form1 was filled
  is($_(1, "uname").value, "testuser", "Checking for filled username");
  is($_(1, "pword").value, "testpass", "Checking for filled password");

  // Test that found-logins observer is notified & got correct data
  is(TestObserver.receivedNotificationFoundLogins, true, "Checking found-logins observer was notified");
  is(TestObserver.dataFoundLogins.get("didntFillReason"), "noAutofillForms", "Checking didntFillReason is noAutofillForms");
  is(SpecialPowers.unwrap(TestObserver.dataFoundLogins.get("usernameField")), $_(1, "uname"), "Checking username field is correct");
  is(SpecialPowers.unwrap(TestObserver.dataFoundLogins.get("passwordField")), $_(1, "pword"), "Checking password field is correct");
  is(TestObserver.dataFoundLogins.get("foundLogins").constructor.name, "Array", "Checking foundLogins is array");
  is(TestObserver.dataFoundLogins.get("foundLogins").length, 1, "Checking foundLogins contains one login");
  ok(TestObserver.dataFoundLogins.get("selectedLogin").QueryInterface(Ci.nsILoginInfo), "Checking selectedLogin is nsILoginInfo");
  ok(TestObserver.dataFoundLogins.get("selectedLogin").equals(TestObserver.dataFoundLogins.get("foundLogins")[0]),
     "Checking selectedLogin is found login");

  // Reset pref (since we assumed it was true to start)
  SpecialPowers.setBoolPref("signon.autofillForms", true);

  // Remove the observer
  SpecialPowers.removeObserver(TestObserver, "passwordmgr-found-form");
  SpecialPowers.removeObserver(TestObserver, "passwordmgr-found-logins");

  SimpleTest.finish();
}


</script>
</pre>
</body>
</html>
